Added mu fix6 maths library for 3D graphics on 8bit machines.

+++ src/lib/libfix6/matrix_process_vectors.c
@@ -1,7 +1,7 @@
 
 #include "_fix6.h"
 
-void matrix_process_vectors(matrix_t* mat, vector_t* v, size_t n, vector_t* o)
+void matrix_process_vectors(matrix_t* mat, vector_t* v, int n, vector_t* o)
 {
     for (; n > 0; v++, o++, n--) {
         o->x = v->x*mat->v[0][0]+v->y*mat->v[0][1]+v->z*mat->v[0][2]+i2f6(mat->v[0][3]);
